+++ src/extensibility/dotnet-extensibility/Extensibility/DataSource/IDataSource.cs
@@ -1,28 +1,17 @@
 ﻿// MIT License
 // Copyright (c) [2024] [nexus-main]
 
+using System.Text.Json;
 using Microsoft.Extensions.Logging;
 using Nexus.DataModel;
 
 namespace Nexus.Extensibility;
 
 /// <summary>
-/// A data source.
+/// A base interface for IDataSource&gt;T&lt;.
 /// </summary>
 public interface IDataSource
 {
-    /// <summary>
-    /// Invoked by Nexus right after construction to provide the context.
-    /// </summary>
-    /// <param name="context">The <paramref name="context"/>.</param>
-    /// <param name="logger">The logger.</param>
-    /// <param name="cancellationToken">A token to cancel the current operation.</param>
-    /// <returns>The task.</returns>
-    Task SetContextAsync(
-        DataSourceContext context,
-        ILogger logger,
-        CancellationToken cancellationToken);
-
     /// <summary>
     /// Gets the catalog registrations that are located under <paramref name="path"/>.
     /// </summary>
@@ -31,7 +20,8 @@ public interface IDataSource
     /// <returns>The catalog identifiers task.</returns>
     Task<CatalogRegistration[]> GetCatalogRegistrationsAsync(
         string path,
-        CancellationToken cancellationToken);
+        CancellationToken cancellationToken
+    );
 
     /// <summary>
     /// Enriches the provided <see cref="ResourceCatalog"/>.
@@ -41,7 +31,8 @@ public interface IDataSource
     /// <returns>The catalog request task.</returns>
     Task<ResourceCatalog> EnrichCatalogAsync(
         ResourceCatalog catalog,
-        CancellationToken cancellationToken);
+        CancellationToken cancellationToken
+    );
 
     /// <summary>
     /// Gets the time range of the <see cref="ResourceCatalog"/>.
@@ -51,7 +42,8 @@ public interface IDataSource
     /// <returns>The time range task.</returns>
     Task<(DateTime Begin, DateTime End)> GetTimeRangeAsync(
         string catalogId,
-        CancellationToken cancellationToken);
+        CancellationToken cancellationToken
+    );
 
     /// <summary>
     /// Gets the availability of the <see cref="ResourceCatalog"/>.
@@ -65,7 +57,8 @@ public interface IDataSource
         string catalogId,
         DateTime begin,
         DateTime end,
-        CancellationToken cancellationToken);
+        CancellationToken cancellationToken
+    );
 
     /// <summary>
     /// Performs a number of read requests.
@@ -83,5 +76,45 @@ public interface IDataSource
         ReadRequest[] requests,
         ReadDataHandler readData,
         IProgress<double> progress,
-        CancellationToken cancellationToken);
+        CancellationToken cancellationToken
+    );
+}
+
+#pragma warning disable CS0618 // Type or member is obsolete
+
+/// <summary>
+/// A data source.
+/// </summary>
+/// <typeparam name="T">The source configuration type</typeparam>
+public interface IDataSource<T> : IDataSource
+{
+    /// <summary>
+    /// Invoked by Nexus right after construction to provide the context.
+    /// </summary>
+    /// <param name="context">The context.</param>
+    /// <param name="logger">The logger.</param>
+    /// <param name="cancellationToken">A token to cancel the current operation.</param>
+    /// <returns>The task.</returns>
+    Task SetContextAsync(
+        DataSourceContext<T> context,
+        ILogger logger,
+        CancellationToken cancellationToken
+    );
+}
+
+#pragma warning restore CS0618 // Type or member is obsolete
+
+/// <summary>
+/// Data sources which have configuration data to be upgraded should implement this interface.
+/// </summary>
+public interface IUpgradableDataSource
+{
+    /// <summary>
+    /// Upgrades the source configuration.
+    /// </summary>
+    /// <param name="configuration">The configuration.</param>
+    /// <returns>The upgraded source configuration.</returns>
+    static abstract Task<JsonElement> UpgradeSourceConfigurationAsync(
+        JsonElement configuration
+    );
 }