+++ tests/Nexus.Tests/Services/PipelineServiceTests.cs
@@ -96,15 +96,59 @@ public class PipelineServiceTests
         var pipelineService = GetPipelineService(default!, userToPipelinesMap);
 
         // Act
-        var actual = pipelineService.TryGet(USERNAME_1, id2, out var actualPipeline);
+        var success = pipelineService.TryGet(USERNAME_1, id2, out var actualPipeline);
 
         // Assert
-        Assert.True(actual);
+        Assert.True(success);
 
         Assert.Equal(
-            expected: JsonSerializer.Serialize(actualPipeline),
-            actual: JsonSerializer.Serialize(userToPipelinesMap[USERNAME_1][id2])
+            expected: JsonSerializer.Serialize(userToPipelinesMap[USERNAME_1][id2]),
+            actual: JsonSerializer.Serialize(actualPipeline)
         );
+    }
+
+    [Fact]
+    public async Task CanTryUpgradePipeline()
+    {
+        // Arrange
+        var id1 = Guid.NewGuid();
+        var id2 = Guid.NewGuid();
+
+        var userToPipelinesMap = new Dictionary<string, Dictionary<Guid, DataSourcePipeline>>()
+        {
+            [USERNAME_1] = new()
+            {
+                [id1] = new DataSourcePipeline(
+                    Registrations: [],
+                    ReleasePattern: ".^"
+                ),
+                [id2] = new DataSourcePipeline(
+                    Registrations: [],
+                    ReleasePattern: ".*"
+                )
+            }
+        };
+
+        var filePath = Path.GetTempFileName();
+        var pipelineService = GetPipelineService(filePath, userToPipelinesMap);
+
+        var newPipeline = new DataSourcePipeline(
+            Registrations: [],
+            ReleasePattern: "foo"
+        );
+
+        var expected = userToPipelinesMap[USERNAME_1];
+        expected[id1] = newPipeline;
+
+        // Act
+        var success = await pipelineService.TryUpdateAsync(USERNAME_1, id1, newPipeline);
+
+        // Assert
+        Assert.True(success);
+
+        var actual = JsonSerializer.Deserialize<Dictionary<Guid, DataSourcePipeline>>(File.ReadAllText(filePath));
+
+        Assert.Equivalent(expected, actual, strict: true);
     }
 
     [Fact]